requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Markdown==3.7
openai>=2.0.0,<3.0.0
pydantic==2.9.2
python-dotenv==1.0.1
